chat_gpt_classification/requirements.txt
dbt-core==1.8.0-b3
dbt-snowflake==1.8.0b3
snowflake-connector-python==3.9.1